Put space-significant expression in single quotes
This is because in Perl6 regexps, spaces are no longer significant.  Thus,
if one wants to match a string containing spaces, either wrap the string in
single quotes or specify the `:s` adverb to tell the regexp engine that
spaces are significant.  I chose here the non-adverb form since now it's no
longer necessary to escape the '.'.